Moves services implementations to E5R.Tools.Bit.Engine component
