+++ src/E5R.Tools.Bit.Engine/Services/BitDiscovery.cs
@@ -1,0 +1,12 @@
+// Copyright (c) E5R Development Team. All rights reserved.
+// Licensed under the Apache License, Version 2.0. More license information in LICENSE.txt.
+
+namespace E5R.Tools.Bit.Engine.Services
+{
+    using Sdk.Bit.Services.Abstractions;
+
+    public class BitDiscovery : IBitDiscovery
+    {
+
+    }
+}
